+++ docs/content/docs/guides/billing.mdx
@@ -42,7 +42,7 @@ Only the workspace owner can change the plan. Other members see the same compari
 
 Need a single limit raised without moving plan? Request an increase from **Settings > Limits**, linked from the usage section. An operator reviews it and the approved limit takes effect on the workspace.
 
-**Plans** compares the lineup, applies promo codes, and lists codes this workspace has redeemed. On a paid plan each card also prices the switch before you commit, showing what is due today and when the next bill lands.
+**Plans** compares the lineup as full plan cards, with the next step up from your current plan highlighted. Below them, a side by side table puts every limit and capability in one grid, with your current plan column highlighted so you can see exactly what would change. Limits shown are the ones this instance enforces. On a paid plan each card also prices the switch before you commit, showing what is due today and when the next bill lands. Promo codes and the codes this workspace has redeemed live here too.
 
 **AI and credits** covers the credit balance, top-ups and usage. See [AI credits](/guides/ai-credits/).
